Criação do menu inicial e atualização do "HomeFragment" para "MenuFragment", assim como as devidas funcionalidades.
