+++ app/src/main/java/com/example/reitoriaitinerante/ui/menu/MenuFragment.java
@@ -1,0 +1,66 @@
+package com.example.reitoriaitinerante.ui.menu;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.example.reitoriaitinerante.databinding.FragmentHomeBinding;
+
+public class MenuFragment extends Fragment {
+
+    private FragmentHomeBinding binding;
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        MenuViewModel menuViewModel =
+                new ViewModelProvider(this).get(MenuViewModel.class);
+
+        binding = FragmentHomeBinding.inflate(getLayoutInflater(), container, false);
+        View root = binding.getRoot();
+
+        final Button adicionarSugestaoButton = binding.adicionarSugestaoButton;
+        final Button verSugestaoButton = binding.verSugestaoButton;
+        final Button progamacaoButton = binding.programacaoButton;
+        final Button estatisticaButton = binding.estatisticaButton;
+        adicionarSugestaoButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent();
+            }
+        });
+
+        verSugestaoButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent();
+            }
+        });
+        progamacaoButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent();
+            }
+        });
+        estatisticaButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent();
+            }
+        });
+
+        return root;
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
+    }
+}
