+++ app/src/main/java/com/example/reitoriaitinerante/ui/menu/MenuViewModel.java
@@ -1,0 +1,9 @@
+package com.example.reitoriaitinerante.ui.menu;
+
+import androidx.lifecycle.ViewModel;
+
+public class MenuViewModel extends ViewModel {
+
+    public MenuViewModel() {
+    }
+}
