+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.example.reitoriaitinerante.ui.home.HomeFragment"
+        android:name="com.example.reitoriaitinerante.ui.menu.MenuFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home" />
