+++ app/src/main/res/values/preloaded_fonts.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <array name="preloaded_fonts" translatable="false">
+        <item>@font/barlow_bold</item>
+    </array>
+</resources>
